better setup for input data checking
