+++ src/Poseidon/Package.hs
@@ -10,7 +10,6 @@ module Poseidon.Package (
     filterDuplicatePackages,
     findAllPoseidonYmlFiles,
     readPoseidonPackageCollection,
-    getChecksum,
     getJointGenotypeData,
     getJointJanno,
     getJointIndividualInfo,
@@ -37,7 +36,8 @@ import           Poseidon.PoseidonVersion   (asVersion, latestPoseidonVersion,
 import           Poseidon.SecondaryTypes    (ContributorSpec (..), IndividualInfo(..))
 import           Poseidon.Utils             (PoseidonException (..),
                                              renderPoseidonException,
-                                             PoseidonLogIO, LogMode (..), usePoseidonLogger)
+                                             PoseidonLogIO, LogMode (..), usePoseidonLogger,
+                                             checkFile)
 
 import           Colog                      (logInfo, logWarning, logDebug)
 import           Control.Exception          (throwIO, try, catch)
@@ -48,9 +48,7 @@ import           Data.Aeson                 (FromJSON, ToJSON, object,
                                              parseJSON, toJSON, withObject,
                                              (.:), (.:?), (.=))
 import qualified Data.ByteString            as B
-import qualified Data.ByteString.Lazy       as LB
 import           Data.Char                  (isSpace)
-import           Data.Digest.Pure.MD5       (md5)
 import           Data.Either                (lefts, rights)
 import           Data.List                  (elemIndex, groupBy, intercalate,
                                              nub, sortOn, (\\))
@@ -73,8 +71,7 @@ import           SequenceFormats.Eigenstrat (EigenstratIndEntry (..), GenoEntry(
                                              readEigenstratSnpFile)
 import           SequenceFormats.Plink      (readBimFile)
 import           System.Console.ANSI        (hClearLine, hSetCursorColumn)
-import           System.Directory           (doesDirectoryExist, doesFileExist,
-                                             listDirectory)
+import           System.Directory           (doesDirectoryExist, listDirectory)
 import           System.FilePath            (takeDirectory, takeExtension,
                                              takeFileName, (</>), takeBaseName)
 import           System.IO                  (IOMode (ReadMode), hFlush,
@@ -394,24 +391,6 @@ checkFiles baseDir ignoreChecksums ignoreGenotypeFilesMissing yml = do
             checkFile (d </> genoFile gd) $ genoFileChkSum gd
             checkFile (d </> snpFile gd) $ snpFileChkSum gd
             checkFile (d </> indFile gd) $ indFileChkSum gd
-
-checkFile :: FilePath -> Maybe String -> IO ()
-checkFile fn maybeChkSum = do
-    fe <- doesFileExist fn
-    if not fe
-    then throwM (PoseidonFileExistenceException fn)
-    else
-        case maybeChkSum of
-            Nothing -> return ()
-            Just chkSum -> do
-                fnChkSum <- getChecksum fn
-                when (fnChkSum /= chkSum) $ throwM (PoseidonFileChecksumException fn)
-
-getChecksum :: FilePath -> IO String
-getChecksum f = do
-    fileContent <- LB.readFile f
-    let md5Digest = md5 fileContent
-    return $ show md5Digest
 
 checkJannoIndConsistency :: String -> [JannoRow] -> [EigenstratIndEntry] -> IO ()
 checkJannoIndConsistency pacName janno indEntries = do
